Mostly-working version of in-memory objects with basic filtering of INFO and INV: YearInRange, KeyWordList (ANY and ALL), Category, INV-WantHave, based on Array.filter() in Swift. There is a bug when switching from INFO to INV due to my implementation of ViewFiltering that needs to access the baseItem object, and that being nil (corrupted?) in that situation. My attempted solutions will involve speeding up access using a dictionary (i.e., index), and then just keeping the base in all cases, rather than trying to optimize. The optimization was attempted because just filtering a few hundred INV items takes several seconds on the simulator (prob faster on device, but still). Since I can't figure out how to get a handle on fixing this, I will pull back in the CoreData implementation and attempt to get that working, which would be useful in any case. It ought to be fine, since I solved the problem I was having with Swift array corruption in the casts before (see func fromNSArray() for solution). This backup precedes that project.

+++ StampCollection/DealerItem.swift
@@ -87,7 +87,7 @@ class DealerItem: NSObject {
         return DealerItem.setDataValuesForObject(DealerItem(), fromData: data)
     }
 
-    static func filterArray( collection: [DealerItem], byCategory category: Int ) -> [DealerItem] {
+    static func filterArray( collection: [DealerItem], byCategory category: Int16 ) -> [DealerItem] {
         if category == CollectionStore.CategoryAll {
             return collection
         }
